+++ src/openpile/utils/py_curves.py
@@ -434,13 +434,12 @@ def reese_weakrock(
 @njit(cache=True)
 def custom_pisa_sand(
     sig: float,
-    X: float,
     G0: float,
     D: float,
-    y_ult:float,
+    X_ult:float,
     n:float,
     k:float,
-    p_ult:float,
+    Y_ult:float,
     output_length: int = 20,
 ):
     """Creates a lateral spring with the PISA sand formulation and custom user inputs.
@@ -449,17 +448,15 @@ def custom_pisa_sand(
     ----------
     sig : float
         vertical/overburden effective stress [unit: kPa]
-    X : float
-        Depth below ground level [unit: m]
     G0 : float
         Small-strain shear modulus [unit: kPa]
-    y_ult : float
+    X_ult : float
         Normalized displacement at maximum strength
     k : float
         Normalized stiffness parameter
     n : float
         Normalized curvature parameter, must be between 0 and 1
-    p_ult : float
+    Y_ult : float
         Normalized maximum strength parameter
     output_length : int, optional
         Number of datapoints in the curve, by default 20
@@ -472,21 +469,20 @@ def custom_pisa_sand(
         p vector [unit: kN/m]
     """
     # calculate normsalised conic function
-    y, p = conic(y_ult, n, k, p_ult, output_length)
+    y, p = conic(X_ult, n, k, Y_ult, output_length)
 
     # return non-normalised curve
     return y * (sig * D / G0), p * (sig * D)
 
 @njit(cache=True)
-def cowden_clay(
-    X: float,
+def custom_pisa_clay(
     Su: float,
     G0: float,
     D: float,
-    y_ult:float,
+    X_ult:float,
     n:float,
     k:float,
-    p_ult:float,
+    Y_ult:float,
     output_length: int = 20,
 ):
     """
@@ -496,21 +492,19 @@ def cowden_clay(
 
     Parameters
     ----------
-    X : float
-        Depth below ground level [unit: m]
     Su : float
         Undrained shear strength [unit: kPa]
     G0 : float
         Small-strain shear modulus [unit: kPa]
     D : float
         Pile diameter [unit: m]
-    y_ult : float
+    X_ult : float
         Normalized displacement at maximum strength
     k : float
         Normalized stiffness parameter
     n : float
         Normalized curvature parameter, must be between 0 and 1
-    p_ult : float
+    Y_ult : float
         Normalized maximum strength parameter
     output_length : int, optional
         Number of datapoints in the curve, by default 20
@@ -524,7 +518,7 @@ def cowden_clay(
 
     """
     # calculate normsalised conic function
-    y, p = conic(y_ult, n, k, p_ult, output_length)
+    y, p = conic(X_ult, n, k, Y_ult, output_length)
 
     # return non-normalised curve
     return y * (Su * D / G0), p * (Su * D)